+++ src/nvim/plugin/languages.lua
@@ -1,4 +1,4 @@
-local vimrc = require('vimrc')
+local vimrc_pack = require('vimrc_pack')
 
 --- @param bufnr integer
 --- @param vim_filetype string
@@ -16,7 +16,7 @@ local function start_treesitter(bufnr, vim_filetype, ts_lang)
   --   "v:lua.require'nvim-treesitter'.indentexpr()"
 end
 
-vim.pack.add({ 'https://github.com/RRethy/nvim-treesitter-endwise' })
+vimrc_pack.add({ { 'https://github.com/RRethy/nvim-treesitter-endwise' } })
 
 vim.api.nvim_create_autocmd('PackChanged', {
   callback = function(ev)
@@ -29,33 +29,34 @@ vim.api.nvim_create_autocmd('PackChanged', {
   end,
 })
 
-vim.pack.add({
+vimrc_pack.add({
   {
-    src = 'https://github.com/nvim-treesitter/nvim-treesitter',
+    'https://github.com/nvim-treesitter/nvim-treesitter',
     version = 'main',
-  },
-})
-vimrc.setup_plugin_lazy(function()
-  local all_parsers = require('nvim-treesitter').get_available()
-  vim.api.nvim_create_autocmd('FileType', {
-    callback = function(args)
-      local bufnr = args.buf
-      local vim_filetype = args.match
+    lazy = true,
+    setup = function()
+      local all_parsers = require('nvim-treesitter').get_available()
+      vim.api.nvim_create_autocmd('FileType', {
+        callback = function(args)
+          local bufnr = args.buf
+          local vim_filetype = args.match
 
-      local ts_lang = vim.treesitter.language.get_lang(vim_filetype)
-      if not ts_lang or not vim.tbl_contains(all_parsers, ts_lang) then
-        return
-      end
-
-      require('nvim-treesitter')
-        .install(ts_lang)
-        :await(function(_, success)
-          if not success then
+          local ts_lang = vim.treesitter.language.get_lang(vim_filetype)
+          if not ts_lang or not vim.tbl_contains(all_parsers, ts_lang) then
             return
           end
 
-          start_treesitter(bufnr, vim_filetype, ts_lang)
-        end)
+          require('nvim-treesitter')
+            .install(ts_lang)
+            :await(function(_, success)
+              if not success then
+                return
+              end
+
+              start_treesitter(bufnr, vim_filetype, ts_lang)
+            end)
+        end,
+      })
     end,
-  })
-end)
+  },
+})
